added environmental variable
