+++ examples/binned/sed.py
@@ -5,7 +5,7 @@ import numpy as np
 import matplotlib.pyplot as plt
 
 from synthesizer.filters import FilterCollection
-from synthesizer.grid_sw import Grid
+from synthesizer.grid import SpectralGrid
 from synthesizer.binned import SFH, ZH, generate_sfzh, SEDGenerator
 from synthesizer.plt import single, single_histxy, mlabel
 from unyt import yr, Myr
@@ -19,7 +19,7 @@ if __name__ == '__main__':
 
     grid_name = 'bpass-v2.2.1_chab100-bin_cloudy-v17.0_logUref-2'
 
-    grid = Grid(grid_name)
+    grid = SpectralGrid(grid_name)
 
     # --- define the parameters of the star formation and metal enrichment histories
     sfh_p = {'duration': 100 * Myr }
